content/media/webaudio/test/test_bug867203.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Crashtest for bug 867203</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SpecialPowers.setBoolPref("media.webaudio.enabled", true);
SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var ctx = new AudioContext();

  var panner1 = ctx.createPanner();
  panner1.setVelocity(1, 1, 1);
  ctx.listener.setVelocity(1, 1, 1);
  (function() {
    ctx.createBufferSource().connect(panner1);
  })();
  SpecialPowers.forceGC();
  SpecialPowers.forceCC();
  ctx.createPanner();

  ok(true, "We did not crash.");
  SpecialPowers.clearUserPref("media.webaudio.enabled");
  SimpleTest.finish();
});


</script>
</pre>
</body>
</html>
